+++ programmable_video/test/mock_platform_interface.dart
@@ -8,8 +8,12 @@ class MockInterface extends ProgrammableVideoPlatform {
   var setNativeDebugWasCalled = false;
   var nativeDebug;
   var setSpeakerPhoneOnWasCalled = false;
-  var speakerPhoneOn = false;
   var getSpeakerPhoneOnWasCalled = false;
+  var setAudioSettingsWasCalled = false;
+  var getAudioSettingsWasCalled = false;
+  var disableAudioSettingsWasCalled = false;
+  var speakerphoneOn = false;
+  var bluetoothOn = false;
   var deviceHasReceiverWasCalled = false;
   var getStatsWasCalled = false;
   var connectToRoomWasCalled = false;
@@ -29,7 +33,7 @@ class MockInterface extends ProgrammableVideoPlatform {
   }
 
   @override
-  Future<void> setNativeDebug(bool native) {
+  Future<void> setNativeDebug(bool native, bool audio) {
     setNativeDebugWasCalled = true;
     nativeDebug = native;
     return Future.delayed(Duration(milliseconds: 1));
@@ -37,7 +41,7 @@ class MockInterface extends ProgrammableVideoPlatform {
 
   @override
   Stream<BaseCameraEvent>? cameraStream() {
-    return Stream<BaseCameraEvent>.periodic(Duration(seconds: 1), (x) => SkipableCameraEvent());
+    return Stream<BaseCameraEvent>.periodic(Duration(seconds: 1), (x) => SkippableCameraEvent());
   }
 
   @override
@@ -48,14 +52,40 @@ class MockInterface extends ProgrammableVideoPlatform {
   @override
   Future<bool> setSpeakerphoneOn(bool on) {
     setSpeakerPhoneOnWasCalled = true;
-    speakerPhoneOn = on;
-    return Future.delayed(Duration(milliseconds: 1), () => speakerPhoneOn);
+    speakerphoneOn = on;
+    return Future.delayed(Duration(milliseconds: 1), () => speakerphoneOn);
   }
 
   @override
   Future<bool> getSpeakerphoneOn() {
     getSpeakerPhoneOnWasCalled = true;
-    return Future.delayed(Duration(milliseconds: 1), () => speakerPhoneOn);
+    return Future.delayed(Duration(milliseconds: 1), () => speakerphoneOn);
+  }
+
+  @override
+  Future setAudioSettings(bool speakerphoneEnabled, bool bluetoothPreferred) {
+    setAudioSettingsWasCalled = true;
+    speakerphoneOn = speakerphoneEnabled;
+    bluetoothOn = bluetoothPreferred;
+    return Future.delayed(Duration(milliseconds: 1), () => null);
+  }
+
+  @override
+  Future<Map<String, dynamic>> getAudioSettings() {
+    getAudioSettingsWasCalled = true;
+    final result = {
+      'speakerphoneEnabled': speakerphoneOn,
+      'bluetoothPreferred': bluetoothOn,
+    };
+    return Future.delayed(Duration(milliseconds: 1), () => result);
+  }
+
+  @override
+  Future disableAudioSettings() {
+    disableAudioSettingsWasCalled = true;
+    speakerphoneOn = false;
+    bluetoothOn = false;
+    return Future.delayed(Duration(milliseconds: 1), () => null);
   }
 
   @override