+++ programmable_video/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 0.11.0+1
+
+- Throw `ActiveCallException` if we cannot activate the `AVAudioSession` (iOS) or get Audio Focus (Android) on `connect`.
+
+## 0.11.0
+
+- Added responsive management of audio settings using `setAudioSettings`.
+- Added optional integration with `ocarina` on Android to allow for integrated audio focus management.
+- Made improvements to `AVAudioEngineDevice`.
+- Deprecated `setSpeakerPhoneOn`.
+- Added audio device notifications to dart layer.
+- **Android** Normalized camera ID when interacting with `CameraManager` to address a crash introduced by the formatting used by `Camera1Enumerator`.
+
+## 0.10.0+1
+
+- Bumped minor versions of dependencies.
+
 ## 0.10.0
 
 - **BREAKING**: Migrated TwilioVideo iOS SDK from v3 to v4.